LEDH runnables implementation
- LEDH_Input();
- LEDH_Cyclic();
- LEDH_Output();
- u32 COMH_GetConfirmedLedStates(void);

Tested if the ac state was 1 and no request to turn off. So it shall be kept 1.
Tested if the ac state was 1 and request was to turn off. It shall be off.
Tested if the ac state was 0 and the request was to turn on. It shall be on.
Tested if the ac state was 0 and no request to turn off. So it shall be kept 1.

re handle if the requested states were updated during the LEDH_Output is running.