Handle stuck in LED output with lots of requests
The LEDH no longer resets the index counter if a new request is received from COMH. Else it will continue executing from the same index till it finishes then resets. The LEDH has a new LEDH_Cyclic() which runs in the main loop to neglect the any unwanted ledh request. The function checks whether the specific index is requested or not, if it is not requested it increments the counter to the next index to give fast switching and avoid delays if the last LED of index 17 is requested only. Else if the counter reaches a specific requested LED index, the counter stops and the LEDH_Output() debounces on the index until it is processed, then increments the counter. The LEDH_Cyclic is handled as long as there are requests. If there is no requests the counter is reset and nothing is executed. 
TO BE TESTED